CSC4350/finalProject/requirements.txt
Flask
Flask_Login
Flask_SQLAlchemy
python-dotenv
requests
Jinja2
pylint
psycopg2-binary
SQLAlchemy
werkzeug
